Finished chapter 4
Was a pain to compile, still using vscode, since I have no idea how
MAKEFILEs work for Java. Also package naming is weird to me, had to use
"java_lox" instead of "java" as the folder name since it showed up in
the package. Unsure how the author got away with this.

As for the hello world code, just threw it away, no use anyway.